Update and add functons
The assignAge function was producing numbers outside the desired range
so updated the mathematical function to only generate values within the
acceptable range. Applied the same principle to the assignBMI function.
Created two other functions to assign location types and income levels.
Used the if else loop to make urban and low income option most likely
and to make rural and high-income levels the least likely to appear as
to ensure the data was more realistic.